Squashed 'airdcpp-core/' changes from 2147dfa97..00813ba88
00813ba88 Use std::optional with MSVC
70a0f2bbf Add a missing check for matching paths for incoming search results
c972d0e53 Make the code compatible with std::optional
353d8140d MSVC: compile core and API with /std:c++17, fix a few compiler warnings

git-subtree-dir: airdcpp-core
git-subtree-split: 00813ba88eaebe5b863a33f86edc874887dfdb2f

+++ airdcpp/modules/AutoSearch.cpp
@@ -360,9 +360,9 @@ bool AutoSearch::updateSearchTime() noexcept {
 		//add the start (or end) hours and minutes (if needed)
 		auto& timeStruct = toEnabled ? startTime : endTime;
 		if (timeStruct.hour > nextSearch.time_of_day().hours()) {
-			nextSearch += (hours(timeStruct.hour) + minutes(timeStruct.minute)) - (hours(nextSearch.time_of_day().hours()) + minutes(nextSearch.time_of_day().minutes()));
+			nextSearch += (hours(timeStruct.hour) + minutes(timeStruct.minute)) - (hours(static_cast<long>(nextSearch.time_of_day().hours())) + minutes(static_cast<long>(nextSearch.time_of_day().minutes())));
 		} else if ((timeStruct.hour == nextSearch.time_of_day().hours() && timeStruct.minute > nextSearch.time_of_day().minutes())) {
-			nextSearch += minutes(timeStruct.minute - nextSearch.time_of_day().minutes());
+			nextSearch += minutes(timeStruct.minute - static_cast<long>(nextSearch.time_of_day().minutes()));
 		}
 	};
